transpose/requirements.txt
kubernetes
pytest
pyyaml
